Funcionalidades de Tarefas feitas, mudanças no IntegranteController

+++ SistemaScrum/src/main/java/com/pds/sistemascrum/controller/IntegranteController.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,6 +18,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.pds.sistemascrum.model.Integrante;
 import com.pds.sistemascrum.repository.IntegranteRepository;
 
+import javax.validation.Valid;
+
 @RestController
 @RequestMapping("/api/integrantes")
 public class IntegranteController {
@@ -24,8 +28,8 @@ public class IntegranteController {
 	public IntegranteRepository integranteReposiroty;
 	
 	@PostMapping("")
-	public Integrante cadastrarIntegrante(@RequestBody Integrante integrante) {
-		return integranteReposiroty.save(integrante);
+	public ResponseEntity<Integrante> cadastrarIntegrante(@Valid @RequestBody Integrante integrante) {
+		return new ResponseEntity<>(integranteReposiroty.save(integrante), HttpStatus.OK);
 	}
 	
 	@GetMapping("")
@@ -34,22 +38,28 @@ public class IntegranteController {
 	}
 	
 	@GetMapping("/{id}")
-	public Optional<Integrante> recuperarIntegrante(@PathVariable Integer id) {
-		return integranteReposiroty.findById(id);
+	public ResponseEntity<Integrante> recuperarIntegrante(@PathVariable Integer id) {
+		if(integranteReposiroty.existsById(id))
+			return new ResponseEntity<>(integranteReposiroty.findById(id).get(), HttpStatus.OK);
+		else
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 	}
 	
 	@DeleteMapping("/{id}")
-	public Integer deletarInteegrante(@PathVariable Integer id) {
-		integranteReposiroty.deleteById(id);
-		return id;
+	public ResponseEntity<Integer> deletarIntegrante(@PathVariable Integer id) {
+		if(integranteReposiroty.existsById(id)) {
+			integranteReposiroty.deleteById(id);
+			return new ResponseEntity<>(id,HttpStatus.OK);
+		}
+		return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 	}
 	
 	@PutMapping("/{id}")
-	public Integrante atualizarIntegrante(@PathVariable Integer id, @RequestBody Integrante integrante) {
+	public ResponseEntity<Integrante> atualizarIntegrante(@PathVariable Integer id, @Valid @RequestBody Integrante integrante) {
 		if(integranteReposiroty.existsById(id)) {
 			integrante.setId(id);
-			return integranteReposiroty.save(integrante);
+			return new ResponseEntity<>(integranteReposiroty.save(integrante), HttpStatus.OK);
 		} else
-			return null;
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 	}
 }
